backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.12
pydantic==2.10.2
pydantic-settings==2.6.1

SQLAlchemy==2.0.36
psycopg[binary]==3.2.3

redis==5.2.0
celery==5.4.0

Pillow==10.4.0
numpy==2.1.3

ultralytics==8.3.31
opencv-python-headless==4.10.0.84
